feat: update MoMo payment request structure and enhance error handling

+++ frontend/app/src/main/java/com/example/my_mobile_app/api/dto/CreateMoMoPaymentRequest.java
@@ -2,14 +2,14 @@ package com.example.my_mobile_app.api.dto;
 
 public class CreateMoMoPaymentRequest {
     public String orderId;
-    public double amount;
+    public long amount;
     public String orderInfo;
     /** captureWallet | payWithMethod | linkAndPay | linkAndPayWithToken. */
     public String requestType;
 
     public CreateMoMoPaymentRequest() {}
 
-    public CreateMoMoPaymentRequest(String orderId, double amount, String orderInfo) {
+    public CreateMoMoPaymentRequest(String orderId, long amount, String orderInfo) {
         this.orderId = orderId;
         this.amount = amount;
         this.orderInfo = orderInfo;
